Only one players programming card are shown. Still needs update on playerID and updateMoves()

+++ roborally/src/main/java/dk/dtu/compute/se/pisd/roborally/client/Client.java
@@ -103,7 +103,7 @@ public class Client {
             e.printStackTrace();
         }
     }
-    public void getMovesByPlayerID(Long gameID, int playerID) {
+    public void getMovesByPlayerID(int gameID, int playerID) {
         try {
             HttpRequest request = HttpRequest.newBuilder()
                     .GET()
